requirements.txt
Django==3.0.2
flake8==3.7.9
django-environ==0.4.5
gunicorn==20.0.4
django-heroku==0.3.1

django-mysql==3.3.0
djangorestframework==3.11.0
